Minichain.ChatModels.Types

Documentation

type Minichain model = StateT model IO

type ApiKey = ByteString

data Role

Way of distinguising who said what in a conversation

Constructors

User

Assistant

System

data Message

A message with a role and content (lenses role and content)

Constructors

Message

Fields

_role :: Role

_content :: String

role :: Lens' Message Role

content :: Lens' Message String

pattern UserMessage :: String -> Message

Pattern synonym for creating a Message with User role

class ChatModel a where

A class for Chat Models

Minimal complete definition

predictsMsgs, addMsgsTo

Methods

predict :: (MonadIO m, MonadThrow m) => String -> StateT a m [Message]

Predict for current and only model

predicts :: (MonadIO m, MonadThrow m) => Lens' s a -> String -> StateT s m [Message]

Predict for a specific model via lens

predictMsg :: (MonadIO m, MonadThrow m) => Message -> StateT a m [Message]

Predict Message type for current and only model

predictsMsg :: (MonadIO m, MonadThrow m) => Lens' s a -> Message -> StateT s m [Message]

Predict Message type for specific model

predictMsgs :: (MonadIO m, MonadThrow m) => [Message] -> StateT a m [Message]

Predict Multiple Message types for current and only model

predictsMsgs :: (MonadIO m, MonadThrow m) => Lens' s a -> [Message] -> StateT s m [Message]

Predict Multiple Message types for specific model

addMsgs :: Monad m => [Message] -> StateT a m ()

Add Messages for current and only model

addMsgsTo :: Monad m => Lens' s a -> [Message] -> StateT s m ()

Add Messages for specific model

module Control.Monad.State